Updating the quantum extension to add simulator support (#3575)
This PR edits the output function to allow for jobs that run on the simulator to have output logged during execution that is not in json format. For simulator jobs, we expect the job result to be printed as the last item on its own line, and string output to be enclosed with `"`.